[Site] Use ux-icons (1/2)
