+++ ux.symfony.com/assets/styles/components/_DocsLink.scss
@@ -43,6 +43,7 @@
   color: gray;
   font-size: 1.5rem;
   opacity: .5;
+  margin-inline-start: auto;
   transition: opacity 600ms ease-in-out;
   .DocsLink:hover & {
     transition: opacity 250ms ease-in-out;
