+++ ux.symfony.com/templates/demos/live_memory/components/LiveMemory/Icon.html.twig
@@ -1,1 +1,1 @@
-{{ rawSvg|raw }}
+{{ this.getRawSvg()|raw }}
